// src/types/ip.ts
export interface IPResponse {
  status?: string
  message?: string
  continent?: string
  continentCode?: string
  country?: string
  countryCode?: string
  region?: string
  regionName?: string
  city?: string
  district?: string
  zip?: string
  lat?: number
  lon?: number
  timezone?: string
  offset?: number
  currency?: string
  isp?: string
  org?: string
  as?: string
  asname?: string
  reverse?: string
  mobile?: boolean
  proxy?: boolean
  hosting?: boolean
  query?: string
  req_left?: number
  resets_in?: number
  ip_type?: string
  ip?: string
}